lambdas/stack-exchange/requirements.txt
py-stackexchange
slackclient